+++ include/arch/cluster/riscv32-cluster/_riscv32-cluster.h
@@ -22,11 +22,11 @@
  * SOFTWARE.
  */
 
-#ifndef _CLUSTER_RISCV32_SMP_H_
-#define _CLUSTER_RISCV32_SMP_H_
+#ifndef _CLUSTER_RISCV32_CLUSTER_H_
+#define _CLUSTER_RISCV32_CLUSTER_H_
 
 	#define __NEED_CORE_RV32I
 	#include <arch/core/rv32i.h>
 
-#endif /* _CLUSTER_RISCV32_SMP_H_ */
+#endif /* _CLUSTER_RISCV32_CLUSTER_H_ */
