+++ src/hal/arch/core/k1b/_core.S
@@ -26,7 +26,7 @@
 #define _ASM_FILE_
 
 #include <arch/core/k1b/asm.h>
-#include <arch/cluster/k1b/memory.h>
+#include <arch/cluster/k1b-cluster/memory.h>
 
 .global _k1b_core_reset
