+++ src/hal/arch/core/k1b/ivt.c
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-#include <arch/cluster/k1b/cores.h>
+#include <arch/cluster/k1b-cluster/cores.h>
 #include <nanvix/const.h>
 
 /**
