+++ src/hal/arch/core/or1k/boot_code.S
@@ -45,7 +45,7 @@
 #include <arch/core/or1k/core.h>
 #include <arch/core/or1k/mmu.h>
 #include <arch/core/or1k/int.h>
-#include <arch/cluster/or1k/memory.h>
+#include <arch/cluster/or1k-cluster/memory.h>
 
 /* Exported symbols. */
 .globl start
